refactor: posts static content 모듈화
